refactor: move share image generation out of route handler
